fix(types): output enums in the minified bundle
