+++ package.json
@@ -75,5 +75,8 @@
         "commitizen": {
             "path": "./node_modules/cz-conventional-changelog"
         }
-    }
+    },
+    "sideEffects": [
+        "./src/resources/Enums.ts"
+    ]
 }
