+++ src/PlatformClient.ts
@@ -1,3 +1,5 @@
+import './resources/Enums';
+
 import API from './APICore';
 import {APIConfiguration, PlatformClientOptions} from './ConfigurationInterfaces';
 import {HostUndefinedError} from './Errors';
