+++ src/resources/BaseInterfaces.ts
@@ -6,15 +6,6 @@ export interface PageModel<T = any> {
 
 export type New<T, K extends string | number | symbol = null> = Omit<T, 'id' | K>;
 
-export enum AuthProvider {
-    SALESFORCE = 'SALESFORCE',
-    SALESFORCE_SANDBOX = 'SALESFORCE_SANDBOX',
-    GOOGLE = 'GOOGLE',
-    OFFICE365 = 'OFFICE365',
-    SAML = 'SAML',
-    EMAIL = 'EMAIL',
-}
-
 export interface IdAndDisplayNameModel {
     id: string;
     displayName?: string;
